pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
dependencies = [
    "python-telegram-bot>=20.7",
    "requests>=2.32.4",
    "flask>=3.1.1"
]
requires-python = ">=3.11"
